Managed to get the YTD and Yearly data to fetch from HealthKit

+++ FitnessApp/Charts/Model/MonthlyStepModel.swift
@@ -10,5 +10,5 @@ import Foundation
 struct MonthlyStepModel : Identifiable {
     let id = UUID()
     let date : Date
-    let count : Double
+    let count : Int
 }
